# Performance enhancement for Dist objects