Remove key function and replace with key attribute.

Since the key is immutable, we don't need a function call to produce it.
Building the key once at object creation avoids a lot of new tuple
objects and function calls.  I also store the hash of the key since
tuples recursively calculate their hash values each time.

In creating a basic data science environment, these changes shave about 1.7s off the total runtime for conda.
```
time python main.py create -n testdev python numpy scipy pandas jupyter --dry-run
dist-key2: 13.32s user 0.20s system 93% cpu 14.521 total

master: 15.33s user 0.21s system 95% cpu 16.223 total
```
The improvement is more noticeable with more packages.
```
time python main.py create -n testdev python numpy scipy pandas jupyter pytables bokeh cython dask datashader holoviews seaborn --dry-run
dist-key2: 30.51s user 0.29s system 96% cpu 32.031 total

master: 35.55s user 0.44s system 96% cpu 37.267 total
```