@groutr 2018-03-03
Looks like the conda-build tests are passing now except for some missing R package from CRAN (r-cran-nmf)