@mbargull 2018-03-03
What's the motive to remove the `assert`s? I'm no big fan of inducing potential silent fails..
@kalefranz: What was the motive that no `assert` is used for `__eq__`?