@groutr 2018-03-05
@mbargull Running/compiling Python code in any optimized mode will disable assert statements, therefore asserts should never be used for control flow. Assertions are debug level statements for checking assumptions in a chunk of code.  We shouldn't assume that Python is always running our code in what is effectively a debug mode.  Using assert statements could introduce the silent fail behavior you wanted to avoid.

As for the silent fail potential, a better way would probably be to return NotImplemented if the type check fails.