@groutr 2018-03-05
I should also explain for the benefit of all why we can't cache the has of self.__key.  What I observed is that the hash was changing between different instances of Dist objects.  For a given package in the solution, the dist object in the solution had a different hash than the dist object in the index.

In one of the failing conda-build tests, we had this situation.  The first line is the object in the solution and the second line is the corresponding object in the index.  The three fields are the objects hash, the contents of the object's key, and the object's id.
```
                   hash                contents of key                    id
Solution: -8194553981133064073 conda-forge ca-certificates-2018.1.18-0 139627647241296
Index: -8752704017545300206 conda-forge ca-certificates-2018.1.18-0 139627677492168

6310709686116748962 conda-forge certifi-2018.1.18-py36_0 139627647241464
4339292024634228224 conda-forge certifi-2018.1.18-py36_0 139627678240160

8068237682851260134 conda-forge ncurses-5.9-10 139627647241128
7380129252822756334 conda-forge ncurses-5.9-10 139627678464600

806033304550791492 conda-forge openssl-1.0.2n-0 139627647241240
-5172741414867245551 conda-forge openssl-1.0.2n-0 139627678617104

6783883222000868768 conda-forge pip-9.0.1-py36_1 139627647239784
4375778942271413089 conda-forge pip-9.0.1-py36_1 139627683574056

-4066167169479619783 conda-forge python-3.6.4-0 139627647239560
5076903470494095772 conda-forge python-3.6.4-0 139627681752008

8612007355177735106 conda-forge readline-7.0-0 139627647240848
-5141768007290159154 conda-forge readline-7.0-0 139627679659568

-1909622333390710995 conda-forge setuptools-38.5.1-py36_0 139627647239448
2847656242649401031 conda-forge setuptools-38.5.1-py36_0 139627681856480

24015384924622500 conda-forge sqlite-3.20.1-2 139627647240736
3893619110724713884 conda-forge sqlite-3.20.1-2 139627717402296

-504215723807823378 conda-forge tk-8.6.7-0 139627647240568
-8546783321497732552 conda-forge tk-8.6.7-0 139627695720936

1287821086695946909 conda-forge wheel-0.30.0-py36_2 139627647240064
2618240198229800662 conda-forge wheel-0.30.0-py36_2 139627678202848

-6611820260125506438 conda-forge xz-5.2.3-0 139627647240344
-8305524842936266159 conda-forge xz-5.2.3-0 139627680903296

-6725206302785262792 conda-forge zlib-1.2.11-0 139627647240232
-4831967795482785717 conda-forge zlib-1.2.11-0 139627698195872
```