@mbargull 2018-03-05
> @mbargull Running/compiling Python code in any optimized mode will disable assert statements, therefore asserts should never be used for control flow. Assertions are debug level statements for checking assumptions in a chunk of code. 

Yeah, I know and I'm also very inclined to rap over people's knuckles if they dare to incorporate `assert` into control flow -- and it's actually one major thing that I dislike about defaults and conventions on how (C)Python is used. I wish `python -OO` were the default, static type checking would become more common, etc. /minor (agreeing) rant over
I assumed (i.e., haven't checked all uses), that those `assert`s were used as they should (i.e., as you depicted, not in control flow and not expected failures) from what I then assumed that the comparisons are never done with non-`Dist` objects. And thus the existing `isinstance` for `__eq__` didn't make sense to me.
(And rather than throwing `NotImplemented`/`TypeError`, I'd make more sense to look review the `Dist` comparison usages -- but I guess, investing time in #6979  and further work makes even more sense :wink:)

----
Thanks for the additional information and investigation regarding the hashes; interesting, I wonder how that happens (something pickled/unpickled? would that make an influence? strange.. ).