@groutr 2018-03-07
@mbargull, I'm glad that we agree about assert statements.
I see what you mean by the silent behavior.  I'm in the process of checking the various versions of these comparison methods to see how they affect the test suite.  I can also gather information on if the type check fails (and what types attempt to be compared with Dist objects).

I haven't seen any code that explicitly compares Dist objects, and the fields that are in the key lead me to believe that the comparisons are probably only used for sorting purposes.  What meaning can you ascribe to the comparison of 'conda-forge' < 'defaults' or comparing the string representation of the dist?

At the end of the day, I agree, if Dist is to go away, then more effort on this PR probably isn't warranted.  I'd much rather put the additional effort into #6979.

As for the hashes, I suspect that somewhere an object is using the default `__hash__` which relies on the object's id.